# docs/swagger.yaml
basePath: /
definitions:
  models.Category:
    properties:
      description:
        example: Aneka Jus dan Kopi
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Minuman
        type: string
    type: object
  models.Product:
    properties:
      categoryId:
        example: 1
        type: integer
      categoryName:
        example: Makanan Berat
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Nasi Goreng Spesial
        type: string
      price:
        example: 25000
        type: integer
      stock:
        example: 100
        type: integer
    type: object
info:
  contact:
    email: intan.maharani6763@gmail.com
    name: Intan Maharani
  description: API Kasir Toko dengan Database (Supabase)
  title: Kasir API Intan
  version: "1.0"
paths:
  /api/products:
    get:
      consumes:
      - application/json
      description: Mengambil daftar semua produk
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Ambil semua produk
      tags:
      - 2. products
    post:
      consumes:
      - application/json
      description: Menambahkan data produk ke database
      parameters:
      - description: Data Produk
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
      summary: Tambah produk baru
      tags:
      - 2. products
  /api/products/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus produk berdasarkan ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hapus produk
      tags:
      - 2. products
    get:
      consumes:
      - application/json
      description: Mendapatkan detail satu produk
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Ambil produk berdasarkan ID
      tags:
      - 2. products
    put:
      consumes:
      - application/json
      description: Mengubah data produk yang sudah ada
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data Update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Update data produk
      tags:
      - 2. products
  /categories:
    get:
      consumes:
      - application/json
      description: Mengambil daftar kategori
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      summary: Ambil semua kategori
      tags:
      - 1. categories
    post:
      consumes:
      - application/json
      description: Membuat kategori baru
      parameters:
      - description: Data Kategori
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Category'
      summary: Tambah kategori
      tags:
      - 1. categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus kategori permanen
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hapus kategori
      tags:
      - 1. categories
    get:
      consumes:
      - application/json
      description: Mendapatkan kategori berdasarkan ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      summary: Ambil detail kategori
      tags:
      - 1. categories
    put:
      consumes:
      - application/json
      description: Mengubah data kategori
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data Update
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      summary: Update kategori
      tags:
      - 1. categories
swagger: "2.0"